% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Cap}
\alias{Cap}
\title{Options for word capitalization}
\usage{
Cap(string, words = c("all", "first"))
}
\arguments{
\item{string}{string of words to capitalize}

\item{words}{whether to capitalize \code{all} words (title case; default) or
only the \code{first}. \code{first} DOES NOT preserve capitalization.}
}
\value{
string of equal length to \code{string} with capitalization applied
}
\description{
Options for word capitalization
}
\examples{
# Title case
Cap("put me in title case please")

# First word only
Cap("only the first word this time.", "first")

# Existing capitalization is NOT preserved
Cap("Since Oklahoma is not the first word, it will not be capitalized.", "first")

# Vectorization is accommodated
Cap(c("title case please", "and me too"))

# Handles missing values
Cap(c("title case please", NA_character_, "and me too"))
}
