% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_itis.R
\name{get_itis}
\alias{get_itis}
\title{Retrieve specific taxon information from ITIS based on scientific name query}
\usage{
get_itis(scientific_names)
}
\arguments{
\item{scientific_names}{string of scientific names with which to query ITIS}
}
\value{
tibble containing the queried scientific name, the valid ITIS scientific
 name if the taxon is of 'Species' rank, the ITIS common name (if present), and
 the ITIS taxon rank of the queried scientific name
}
\description{
Retrieve specific taxon information from ITIS based on scientific name query
}
\examples{
\dontrun{
sn <- c("Thryothorus ludovicianus", "Cardinalis cardinalis", "Bidens alba",
        "Pachydiplax longipennis", "Cladonia evansii", "Ischnura hastata",
        "Smilax bona-nox", "Plantae", "Malvaviscus", "Anaxyrus", "Gulo gulo")
get_itis(sn)
}
}
