% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_vernac}
\alias{get_vernac}
\title{Extract vernacular}
\usage{
get_vernac(string)
}
\arguments{
\item{string}{string returned in 'vernacular' field of
\code{\link[solrium]{solr_search}} to \url{http://services.itis.gov/}. It
is a very specific format...}
}
\value{
string of equal length to \code{string} returning the first vernacular
 English name, in all lower case
}
\description{
Extract vernacular
}
