% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_missing}
\alias{is_missing}
\title{Check for various forms of missing strings beyond NA and ""}
\usage{
is_missing(string)
}
\arguments{
\item{string}{string to evaluate for missing content. Missing content is defined
as NA_character_, empty strings, and strings comprised entirely of white space}
}
\value{
logical of equal length to \code{string} indicating whether content is
 missing
}
\description{
Check for various forms of missing strings beyond NA and ""
}
\examples{
is_missing(c("", "ok", "  ", "       ", NA_character_, "ok, too"))
}
