% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_class}
\alias{retrieve_class}
\title{Extract taxonomic class}
\usage{
retrieve_class(string)
}
\arguments{
\item{string}{string returned in 'hierarchySoFarWRanks' field of
\code{\link[solrium]{solr_search}} to \url{http://services.itis.gov/}. It
is a very specific format...}
}
\value{
string of equal length to \code{string} returning the ITIS
 taxonomic class, if present
}
\description{
Extract taxonomic class
}
