% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_rank}
\alias{retrieve_rank}
\title{Extract taxon rank}
\usage{
retrieve_rank(string)
}
\arguments{
\item{string}{string returned in 'hierarchySoFarWRanks' field of
\code{\link[solrium]{solr_search}} to \url{http://services.itis.gov/}. It
is a very specific format...}
}
\value{
string of equal length to \code{string} returning the lowest taxonomic
 rank associated with that taxon
}
\description{
Extract taxon rank
}
