% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_sci_name}
\alias{retrieve_sci_name}
\title{Extract valid scientific name, if present rank}
\usage{
retrieve_sci_name(string)
}
\arguments{
\item{string}{string returned in 'hierarchySoFarWRanks' field of
\code{\link[solrium]{solr_search}} to \url{http://services.itis.gov/}. It
is a very specific format...}
}
\value{
string of equal length to \code{string} returning the valid ITIS
 scientific name, if present
}
\description{
Extract valid scientific name, if present rank
}
